# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TERRAST Sustainability Disclosure Hub Application API
  version: 0.1.0-concept
  summary: Internal application command/query contract for the concept MVP
  description: |
    Design contract for server-side application endpoints. It is not a TERRAST API
    specification and does not assert that these routes are deployed. The production
    TERRAST endpoint, authentication, and payload contract are unknown; no external
    TERRAST path appears in this document. Implementations must authenticate, enforce
    tenant membership/consent server-side, validate with Zod, and preserve RLS.
  contact:
    name: Repository maintainers
servers:
  - url: http://localhost:3000
    description: Local development only
...
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /api/ai/disclosure:
    post:
      tags: [AI]
      operationId: generateDisclosureDraft
      summary: Generate a structured, evidence-linked disclosure proposal
      description: |
        Server-side only. Returns insufficient_evidence instead of unsupported facts.
        Every successful draft remains an AI proposal requiring human review. The
        x-demo-role header is accepted only while explicit demo mode is enabled.
      x-implementation-status: implemented
      security:
        - bearerAuth: []
        - demoRole: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiDraftRequest"
      responses:
        "200":
          description: Validated AI or deterministic-demo result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiDraftResponse"
        "400":
          description: Request body failed strict validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "401":
          description: Missing/invalid identity or disallowed role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "403":
          description: The authenticated principal cannot access one or more requested sources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "429":
          description: In-process demo request budget exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "503":
          description: Authentication, source access, server configuration, or atomic audit persistence is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "502":
          description: Provider generation/validation failed; deterministic fallback is available
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [error, correlationId, fallbackAvailable]
                properties:
                  error:
                    type: string
                    const: ai_generation_failed
                  correlationId:
                    type: string
                  fallbackAvailable:
                    type: boolean
                    const: true
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Illustrative session name; implementation must use its configured secure cookie contract.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    demoRole:
      type: apiKey
      in: header
      name: x-demo-role
      description: Explicit demo-mode identity only; never a production authorization mechanism.
...
  schemas:
...
    AiDraftRequest:
      type: object
      additionalProperties: false
      required:
        [
          companyId,
          requirementId,
          requirementSummary,
          sourceData,
          evidenceIds,
          requestedTasks,
        ]
      properties:
        companyId:
          type: string
          minLength: 1
          maxLength: 100
        requirementId:
          type: string
          minLength: 1
          maxLength: 100
        requirementSummary:
          type: string
          minLength: 1
          maxLength: 1000
        sourceData:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/AiSourceDatum"
        priorYearData:
          type: array
          default: []
          maxItems: 50
          items:
            $ref: "#/components/schemas/AiSourceDatum"
        evidenceIds:
          type: array
          maxItems: 50
          items:
            type: string
            minLength: 1
            maxLength: 200
        requestedTasks:
          type: array
          maxItems: 10
          items:
            type: string
            minLength: 1
            maxLength: 200
        priorYearText:
          type: string
          maxLength: 5000
    AiSourceDatum:
      type: object
      additionalProperties: false
      required: [id, label, value, unit, period, confidence]
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 200
        label:
          type: string
          minLength: 1
          maxLength: 200
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
        unit:
          type: string
          maxLength: 40
        period:
          type: string
          maxLength: 40
        confidence:
          type: number
          minimum: 0
          maximum: 100
        organizationalBoundary:
          type: string
          maxLength: 200
    AiDraftResponse:
      type: object
      additionalProperties: false
      required: [output, meta]
      properties:
        output:
          $ref: "#/components/schemas/AiDisclosureOutput"
        meta:
          type: object
          additionalProperties: false
          required:
            [promptVersion, inputHash, executedBy, executedAt, mode, model]
          properties:
            promptVersion:
              type: string
            inputHash:
              type: string
              pattern: "^[a-f0-9]{64}$"
            executedBy:
              type: string
            executedAt:
              type: string
              format: date-time
            mode:
              type: string
              enum: [demo, api]
            model:
              type: string
    AiDisclosureOutput:
      type: object
      additionalProperties: false
      required:
        - status
        - label
        - reviewRequired
        - disclosureDraft
        - missingQuestions
        - inconsistencies
        - climateCandidates
        - transitionPlanDraft
        - priorYearChanges
        - weakEvidenceWarnings
        - claims
        - insufficientEvidence
      properties:
        status:
          type: string
          enum: [draft, insufficient_evidence]
        label:
          type: string
          const: AI提案・要レビュー
        reviewRequired:
          type: boolean
          const: true
        disclosureDraft:
          oneOf:
            - $ref: "#/components/schemas/GroundedText"
            - type: "null"
        missingQuestions:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [question, relatedRequirementId, requiredEvidence]
            properties:
              question: { type: string }
              relatedRequirementId: { type: string }
              requiredEvidence: { type: string }
        inconsistencies:
          type: array
          items:
            type: object
            additionalProperties: true
        climateCandidates:
          type: array
          items:
            type: object
            additionalProperties: true
        transitionPlanDraft:
          oneOf:
            - $ref: "#/components/schemas/GroundedText"
            - type: "null"
        priorYearChanges:
          type: array
          items:
            type: object
            additionalProperties: true
        weakEvidenceWarnings:
          type: array
          items:
            type: object
            additionalProperties: true
        claims:
          type: array
          items:
            $ref: "#/components/schemas/GroundedText"
        insufficientEvidence:
          oneOf:
            - type: object
              additionalProperties: false
              required: [reason, missingDataDescriptions]
              properties:
                reason: { type: string }
                missingDataDescriptions:
                  type: array
                  minItems: 1
                  items: { type: string }
            - type: "null"
    GroundedText:
      type: object
      additionalProperties: false
      required: [text, evidenceDataIds]
      properties:
        text:
          type: string
        evidenceDataIds:
          type: array
          items:
            type: string
    SimpleError:
      type: object
      additionalProperties: false
      required: [error, correlationId]
      properties:
        error:
          type: string
        correlationId:
          type: string
...